Organize classes into packages
For now there are the following packages:
- duke
- duke.task
- duke.command
