+++ src/duke/Duke.java
@@ -1,3 +1,6 @@
+package duke;
+
+import duke.command.Command;
 import java.util.Scanner;
 
 public class Duke {
