+++ src/duke/TaskList.java
@@ -1,3 +1,6 @@
+package duke;
+
+import duke.task.Task;
 import java.util.List;
 import java.util.ArrayList;
