+++ src/duke/command/AddDeadlineCommand.java
@@ -1,7 +1,15 @@
-public class AddEventCommand extends Command {
+package duke.command;
+
+import duke.TaskList;
+import duke.Ui;
+import duke.Storage;
+import duke.task.Task;
+import duke.task.Deadline;
+
+public class AddDeadlineCommand extends Command {
     private String arguments;
 
-    public AddEventCommand(String arguments) throws Exception {
+    public AddDeadlineCommand(String arguments) throws Exception {
         if (arguments.length() == 0) {
             throw new Exception("Command `deadline` requires arguments");
         }
@@ -9,11 +17,11 @@ public class AddEventCommand extends Command {
     }
 
     public void execute(TaskList taskList, Ui ui, Storage storage) throws Exception {
-        Task event = Event.fromInput(arguments);
-        taskList.addTask(event);
+        Task deadline = Deadline.fromInput(arguments);
+        taskList.addTask(deadline);
 
         storage.saveTasks(taskList);
-        ui.printMessage("Added the following task:\n    " + event.toString() + "\n" + "You now have "
+        ui.printMessage("Added the following task:\n    " + deadline.toString() + "\n" + "You now have "
                 + taskList.size() + " tasks in your list.");
     }
