+++ src/duke/command/ByeCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.TaskList;
+import duke.Ui;
+import duke.Storage;
+
 public class ByeCommand extends Command {
     public ByeCommand(String arguments) throws Exception {
         if (arguments.length() > 0) {
