+++ src/duke/command/CompleteTaskCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.TaskList;
+import duke.Ui;
+import duke.Storage;
+import duke.task.Task;
+
 public class CompleteTaskCommand extends Command {
     private String arguments;
