+++ src/duke/task/Event.java
@@ -1,3 +1,6 @@
+package duke.task;
+
+import duke.DateTime;
 import java.time.LocalDateTime;
 
 public class Event extends Task {
@@ -11,6 +14,10 @@ public class Event extends Task {
     public Event(String description, LocalDateTime time, boolean isCompleted) {
         super(description, isCompleted);
         this.time = time;
+    }
+
+    public LocalDateTime getTime() {
+        return this.time;
     }
 
     public static Event fromInput(String input) throws Exception {
