--- src/main/java/Command.java
@@ -1,4 +1,0 @@
-abstract class Command {
-    abstract void execute(TaskList taskList, Ui ui, Storage storage) throws Exception;
-    abstract boolean shouldExit();
-}
